+++ Kernel/Modules/AgentTicketActionCommon.pm
@@ -2453,12 +2453,20 @@ sub _Mask {
         if ( $Config->{InformAgent} ) {
 
             # get inform user list
-            my @InformUserID = $ParamObject->GetArray( Param => 'InformUserID' );
+            my %InformAgents;
+            my @InformUserID    = $ParamObject->GetArray( Param => 'InformUserID' );
+            my %InformAgentList = $GroupObject->PermissionGroupGet(
+                GroupID => $GID,
+                Type    => 'ro',
+            );
+            for my $UserID ( sort keys %InformAgentList ) {
+                $InformAgents{$UserID} = $AllGroupsMembers{$UserID};
+            }
 
             if ( $Self->{ReplyToArticle} ) {
 
                 # get email address of all users and compare to replyto-addresses
-                for my $UserID ( sort keys %ShownUsers ) {
+                for my $UserID ( sort keys %InformAgents ) {
                     if ( $ReplyToUserIDs{$UserID} ) {
                         push @InformUserID, $UserID;
                         delete $ReplyToUserIDs{$UserID};
@@ -2469,7 +2477,7 @@ sub _Mask {
             my $InformAgentSize = $ConfigObject->Get('Ticket::Frontend::InformAgentMaxSize')
                 || 3;
             $Param{OptionStrg} = $LayoutObject->BuildSelection(
-                Data       => \%ShownUsers,
+                Data       => \%InformAgents,
                 SelectedID => \@InformUserID,
                 Name       => 'InformUserID',
                 Class      => 'Modernize',